Ensure PSCustomObjects correctly have their enums converted to strings
StoreBroker uses custom Enumerations in order to prevent typos for certain
values.  A side-effect of that, is that we have to do extra work when
converting objects to JSON because ConvertTo-Json will by default
convert an enumeration to its _numerical value_ instead of its _name_.

That's why we use Get-JsonBody, which internally uses ConvertTo-Json after
calling Convert-EnumToString to ensure that all enums have been converted
to their string names first.

The problem we had here was that if the object being passed in was a
PSCustomObject, Convert-EnumToString was leaving everything as-is.
This became a problem in a scenario like New-Submission, which calls
Get-SubmissionRollout, modifies the `state` property, and then calls
Set-SubmissionRollout with that modified object.  That modified object
is a PSCustomObject, but its state is now an Enum value which needs
to be converted back into a string.

There are two potential fixes here:
 1. For all known scenarios where this might occur, append `.ToString()`
    to the enum assignment.  This leaves open the possibility that some
	existing cases might be missed, or that new cases may be introduced
	in the future.
 2. Update Convert-EnumToString to understand PSCustomObjects and do
    the right thing.

This change takes the second approach.