Fix character escaping issues during config generation
Many years ago, #24 attempted to properly escape special characters
in a string by creating its own `Get-EscapedJsonValue` function, but
that only escaped `\`, `\t`, `\r`, `\n`.  That doesn't cover all the
possible characters needing to be encoded though.

At the time, this clearly seemed like a _good_ idea, but looking back
at this code with more experience, it's not clear to me why I didn't
just use `ConvertTo-Json` directly. Doing that now.

One additional update that needed to be done however was to "escape"
in `tag` and `notesForCertification` with '//' with '\\'.  This isn't
_completely_ correct, but not doing so meant that anything after that
would be improperly flagged as a comment by `Remove-Comment` and then
removed.