Port Write-InvocationLog from v2 branch
This ports 48a8791a2f62e57da234427410cb40b2f026b0ed from the v2
branch so that v1 can benefit from normalized invocation logging as well.

This will make future support queries easier to handle, as we'll know
exactly what input that was provided as well as the version of
the module actively in use.