Only resolve MediaRootPath if it's necessary
When `Update-Listing` calls into `Update-ListingImage` to remove
unneeded images, it calls that with `-RemoveOnly` and doesn't bother
passing in a `MediaRootPath`.  However, we still tried to resolve
the `MediaRootPath`, and that could cause an exception since the
path being resolved was blank.

This updates `Update-ListImage` to only resolve `MediaRootPath`
if we're going to use it (meaning that `RemoveOnly` wasn't specified).