Fix telemetry failure for long-running sessions
The telemetry code was creating a singleton base telemetry event the
first time that telemetry was being sent.  That base event included the
date.  If a PowerShell session lasted multiple days, the telemetry
reporting would eventually start to fail because the date being reported
for the event was too old.

We now just create a new telemetry event every time one is requested
(this also ensures that it honors changes to the `ApplicationInsightsKey`),
and now we only cache a GUID that we use to represent the "session" ID
which we consider to be the duration of the current PowerShell session.